requirements.txt
# sudo apt install tesseract-ocr
transformers==4.32.1
torch
Pillow
pytesseract
